Replace zsvjmp.S assembler with an inline C implementation
This makes IRAF more portable; however it also binds the use on that
specific f2c implementation (resp. f2c.h). With this change, it will
be impossible to use a real Fortran compiler instead of the builtin
f2c package.

The "mem" common block is also no longer defined.